+++ front-page.php
@@ -29,24 +29,27 @@
 
 				
 				<section class="pad-top">
-					<div class="pull-l-1of12 pull-r-1of12 cf">
+					<div class="pull-l-1of12 pull-r-1of12 m-padding cf">
 						<div class="uvp">
 							<p>From landlords and tenants to buyers and sellers, we provide consultative commercial real estate solutions focused on building <strong>long-term relationships</strong> with you and <strong>long-term success</strong> for your business.</p>
 						</div>
-						<div class="selling-point m-all t-1of3 d-1of3">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-book-knowledge.svg" alt="Real Estate Knowledge">
-							<h5>A wealth of knowledge.</h5>
-							<p>With over 600 years of combined experience, our team of experts provides you with the latest data, marketing strategies and support you deserve.</p>
-						</div>
-						<div class="selling-point m-all t-1of3 d-1of3">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-us-map-love-indy.svg" alt="Real Estate Knowledge">
-							<h5>Locally-owned.</h5>
-							<p>As the largest, locally-owned firm in Indianapolis, we are deeply-connected experts who care about the success of our city. We are proud of our roots!</p>
-						</div>
-						<div class="selling-point m-all t-1of3 d-1of3 right">
-							<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-focused-on-you.svg" alt="Real Estate Knowledge">
-							<h5>Free to focus on you.</h5>
-							<p>We don&#39;t answer to stockholders. We aren&#39;t an oversized corporation that only cares about boosting our numbers. We care about your numbers.</p>
+
+						<div class="selling-point-grid">
+							<div class="selling-point">
+								<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-book-knowledge.svg" alt="Real Estate Knowledge">
+								<h5>A wealth of knowledge.</h5>
+								<p>With over 600 years of combined experience, our team of experts provides you with the latest data, marketing strategies and support you deserve.</p>
+							</div>
+							<div class="selling-point">
+								<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-us-map-love-indy.svg" alt="Real Estate Knowledge">
+								<h5>Locally-owned.</h5>
+								<p>As the largest, locally-owned firm in Indianapolis, we are deeply-connected experts who care about the success of our city. We are proud of our roots!</p>
+							</div>
+							<div class="selling-point">
+								<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-focused-on-you.svg" alt="Real Estate Knowledge">
+								<h5>Free to focus on you.</h5>
+								<p>We don&#39;t answer to stockholders. We aren&#39;t an oversized corporation that only cares about boosting our numbers. We care about your numbers.</p>
+							</div>
 						</div>
 					</div>
 				</section>