+++ library/partials/sectionCareers.php
@@ -1,44 +1,39 @@
-<section class="pad-top">
-	<div class="pull-l-1of12 pull-r-1of12 m-padding cf">
-		<h2 class="header-dark text-center">Everything you need to succeed.</h2>
-		
-	<!-- Coffee. -->
-		<div class="selling-point m-all t-1of2 d-1of3">
+<section class="pull-l-1of12 pull-r-1of12 m-padding cf">
+	<h2 class="header-dark header-margin text-center">Everything you need to succeed.</h2>
+
+	<div class="selling-point-grid">
+
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-coffee-cup.svg" alt="Coffee Icon">
 			<h5>Unlimited Coffee.</h5>
 			<p>The lifeblood of American business place, we have fresh-brewed coffee every morning from the local Copper Moon Cafe. (It’s recommended to keep your coffee intake limited to 3 cups per day.)</p>
 		</div>
 
-	<!-- Lunch + Learn.-->
-		<div class="selling-point m-all t-1of2 d-1of3">
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-sandwich.svg" alt="Lunch and Learn Icon">
 			<h5>Lunch + Learn.</h5>
 			<p>Learn and expand your career with a tasty catered lunch every other week at our various opportunities to learn something new about marketing yourself, providing better service, and honing your craft.</p>
 		</div>
 
-	<!-- Cool Tools. -->
-		<div class="selling-point m-all t-1of2 d-1of3">
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-computer.svg" alt="Tech and tools icon">
 			<h5>Cool Tools.</h5>
 			<p>Get your own laptop, the Microsoft Office 365 Suite, use our iPad’s for tours, and access your files from anywhere on our VPN. Technical difficulty? We’ve got you covered with our in-house IT support.</p>
 		</div>
 	
-	<!-- Build Business. -->
-		<div class="selling-point m-all t-1of2 d-1of3">
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-sprouting-plant.svg" alt="Networking and Build Business">
 			<h5>Build Business.</h5>
 			<p>Gain access to our paid networking memberships and business groups where you can meet and build relationships with professionals all across Indianapolis, Indiana and beyond.</p>
 		</div>
 
-	<!-- In-house Printing. -->
-		<div class="selling-point m-all t-1of2 d-1of3">
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-printer.svg" alt="Real Estate Knowledge">
 			<h5>In-house Printing.</h5>
 			<p>We offer inhouse printing for our brokers, to help them keep costs down and get things done faster. Print posters on our plotter or scan full-sized floor plans. We also can laminate brochures, mailers and more. </p>
 		</div>
 
-	<!-- Medical -->
-		<div class="selling-point m-all t-1of2 d-1of3">
+		<div class="selling-point">
 			<img src="<?php echo get_template_directory_uri(); ?>/library/images/icons/icon-heart.svg" alt="Real Estate Knowledge">
 			<h5>Medical.</h5>
 			<p>By offering a comprehensive plan through Anthem Blue Cross Blue Shield, we give employees access to one of the largest networks in Indiana. Our plan includes an HDHP with a Health Savings Account.</p>