+++ library/partials/sectionUVP.php
@@ -1,20 +1,21 @@
 <section class="pad-top">
-	<div class="pull-l-1of12 pull-r-1of12 cf">
+	<div class="pull-l-1of12 pull-r-1of12 m-padding cf">
 		<div class="uvp">
 			<?php echo types_render_field( "service-main-description" ); ?>
 		</div>
-
-		<div class="selling-point m-all t-1of3 d-1of3">
-			<h5><?php echo types_render_field( "feature-1-headline" ); ?></h5>
-			<p><?php echo types_render_field( "feature-1" ); ?></p>
-		</div>
-		<div class="selling-point m-all t-1of3 d-1of3">
-			<h5><?php echo types_render_field( "feature-2-headline" ); ?></h5>
-			<p><?php echo types_render_field( "feature-2" ); ?></p>
-		</div>
-		<div class="selling-point m-all t-1of3 d-1of3 right">
-			<h5><?php echo types_render_field( "feature-3-headline" ); ?></h5>
-			<p><?php echo types_render_field( "feature-3" ); ?></p>
+		<div class="selling-point-grid">
+			<div class="selling-point">
+				<h5><?php echo types_render_field( "feature-1-headline" ); ?></h5>
+				<p><?php echo types_render_field( "feature-1" ); ?></p>
+			</div>
+			<div class="selling-point">
+				<h5><?php echo types_render_field( "feature-2-headline" ); ?></h5>
+				<p><?php echo types_render_field( "feature-2" ); ?></p>
+			</div>
+			<div class="selling-point">
+				<h5><?php echo types_render_field( "feature-3-headline" ); ?></h5>
+				<p><?php echo types_render_field( "feature-3" ); ?></p>
+			</div>
 		</div>
 	</div>
 </section>