+++ post-formats/content-post.php
@@ -5,13 +5,11 @@
 		$icon = types_render_field( "report-icon", array( 'raw' => true) );
 	} ?>
 
-<div class="m-all t-1of2 d-1of3">
 	<div class="service-box">
 		<img src="<?php echo $icon; ?>" alt="Specialty">
 
-		<h4><a href="<?php esc_url( the_permalink() ); ?>" rel="bookmark" title="<?php the_title_attribute(); ?>"><?php the_title(); ?></a></h3>
+		<h4><?php the_title(); ?></h3>
 		<?php // wp_trim_excerpt( the_excerpt() ); ?>
 
 		<a class="blue-link" href="<?php esc_url( the_permalink() ); ?>">Read More...</a>
 	</div>
-</div>